Fixed web configuration, Standardized package names, #251
