+++ src/main/java/com/jvm_bloggers/core/rss/converters/SyndFeedXmlMessageConverter.java
@@ -1,4 +1,4 @@
-package com.jvm_bloggers.core.rss.converter;
+package com.jvm_bloggers.core.rss.converters;
 
 import com.rometools.rome.feed.synd.SyndFeed;
 import com.rometools.rome.io.SyndFeedOutput;
