+++ src/test/groovy/com/jvm_bloggers/core/rss/BlogPostsControllerSpec.groovy
@@ -1,14 +1,10 @@
 package com.jvm_bloggers.core.rss
 
 import com.jvm_bloggers.SpringContextAwareSpecification
-import com.rometools.rome.feed.synd.SyndFeed
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.test.web.servlet.MockMvc
 import org.springframework.web.context.WebApplicationContext
-import spock.lang.Ignore
 import spock.lang.Unroll
-
-import javax.servlet.http.HttpServletRequest
 
 import static org.springframework.http.MediaType.APPLICATION_ATOM_XML_VALUE
 import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE
@@ -37,32 +33,6 @@ class BlogPostsControllerSpec extends SpringContextAwareSpecification {
         format || mediaType
         "json" || APPLICATION_JSON_UTF8_VALUE
         "xml"  || APPLICATION_ATOM_XML_VALUE
-        ""     || APPLICATION_JSON_UTF8_VALUE
-    }
-
-    @Ignore
-    def "Should get a raw RSS feed"() {
-        given:
-        HttpServletRequest request = Stub() {
-            getRequestURL() >> new StringBuffer("http://jvm-bloggers.com/rss")
-        }
-        and:
-        AggregatedRssFeedProducer rssProducerMock = Mock() {
-            SyndFeed feed = new TestSyndFeedProvider().getSyndFeed()
-
-            1 * getRss(_, _, _) >> feed
-        }
-        and:
-        BlogPostsController blogPostsController = new BlogPostsController(rssProducerMock, 50)
-
-        when:
-        def result = blogPostsController.getRss(request, null, null)
-
-        then:
-        result != null
-        and:
-        with(result) {
-            getEntries().size() == 1
-        }
+        ""     || APPLICATION_ATOM_XML_VALUE
     }
 }
